move fixtures to the correct conftest
